host close room working
